+++ src/mflux/controlnet/flux_controlnet.py
@@ -36,7 +36,8 @@ class Flux1Controlnet(nn.Module):
         local_path: str | None = None,
         lora_paths: list[str] | None = None,
         lora_scales: list[float] | None = None,
-        controlnet_path: str | None = None,
+        lora_names: list[str] | None = None,
+        lora_repo_id: str | None = None,
     ):
         super().__init__()
         FluxInitializer.init_controlnet(
@@ -46,6 +47,8 @@ class Flux1Controlnet(nn.Module):
             local_path=local_path,
             lora_paths=lora_paths,
             lora_scales=lora_scales,
+            lora_names=lora_names,
+            lora_repo_id=lora_repo_id,
         )
 
     def generate_image(